Refactor dungeon generator components for code organization and consistency (#69)
* Initial plan

* Fix PR review feedback: improve code organization and consistency


---------
